refactor: 타 유저 follow 관련 기능 myPageController로 이동
